+++ src/lib/djerba/plugins/demo1/test/plugin_test.py
@@ -13,7 +13,7 @@ class TestDemo1(PluginTester):
         params = {
             self.INI: 'demo_1.ini',
             self.JSON: 'demo_1.json',
-            self.MD5: 'e16d5ee0888efb70602d476ab708896e'
+            self.MD5: '49199ee9117f159190e0d2ef41beff2e'
         }
         self.run_basic_test(test_source_dir, params)