+++ src/lib/djerba/plugins/sample/test/plugin_test.py
@@ -20,7 +20,7 @@ class TestWgtsSamplePlugin(PluginTester):
         self.maxDiff = None
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
-        sup_dir_var = 'DJERBA_TEST_DATA'
+        sup_dir_var = 'DJERBA_TEST_DIR'
         self.sup_dir = os.environ.get(sup_dir_var)
         
     def testWgtsSample(self):
@@ -31,9 +31,25 @@ class TestWgtsSamplePlugin(PluginTester):
         params = {
             self.INI: ini_location,
             self.JSON: json_location,
-            self.MD5: '3000449df1bac887f778278221776e03'
+            self.MD5: '55a4c4b25376d02f0ca9965afbc63e3f'
         }
         self.run_basic_test(test_source_dir, params)
+
+    def testWgtsSampleWithNA(self):
+        """
+        Purity, ploidy, callability, and coverage are NA
+        """
+        test_source_dir = os.path.realpath(os.path.dirname(__file__))
+        json_location = os.path.join(self.sup_dir ,"plugins/sample/report_json/sample_na.json")
+        ini_location = os.path.join(self.sup_dir ,"plugins/sample/sample_na.ini")
+
+        params = {
+            self.INI: ini_location,
+            self.JSON: json_location,
+            self.MD5: '2a304e1264d51e54497892d8e9f21b2b'
+        }
+        self.run_basic_test(test_source_dir, params)
+
 
 if __name__ == '__main__':
     unittest.main()