+++ src/lib/djerba/plugins/summary/test/plugin_test.py
@@ -24,7 +24,7 @@ class TestSummaryPlugin(PluginTester):
         params = {
             self.INI: 'summary.ini',
             self.JSON: json_location,
-            self.MD5: '3cfe9249b9729fddc02c4a16ebf2e5df'
+            self.MD5: '1c992e2411ecf7b47a6683c9e77194e1'
         }
         self.run_basic_test(test_source_dir, params)